added list of seats that have not been reserved

+++ src/db/migrations/00007_create_showtime.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS showtime (
 	start_time TIMESTAMPTZ NOT NULL,
 	-- This contain both the date and time of the show
 	end_time TIMESTAMPTZ NOT NULL,
-	price INTEGER NOT NULL,
+	price BIGINT NOT NULL,
 	-- pending if the showtime is not active, active if the showtime is active, done if the showtime is over, cancelled if the showtime is cancelled
 	status TEXT DEFAULT 'pending' NOT NULL CHECK (status IN ('pending', 'active', 'done', 'cancelled')),
 	available_seats INTEGER NOT NULL,
